requirements.txt
imageio==2.9.0
loguru==0.5.3
numpy==1.19.5
pandas==1.3.4
Pillow==9.0.1
pydicom==2.2.2
scikit_image==0.18.3
scikit_learn==1.0.2
skimage==0.0
torch==1.10.0
torchvision==0.11.1+cu113
tqdm==4.62.3
wandb==0.12.5
